+++ internal/murmur/murmur_appengine.go
@@ -1,10 +1,10 @@
 // +build appengine
 
-package gocql
+package murmur
 
 import "encoding/binary"
 
-func murmur3H1(data []byte) uint64 {
+func Murmur3H1(data []byte) uint64 {
 	length := len(data)
 
 	var h1, h2, k1, k2 uint64
